Added BurialManager and moved ScoreRotamerVsTarget to own file
