+++ apps/rosetta/riflib/rifdock_tasks/OutputResultsTasks.cc
@@ -13,6 +13,7 @@
 #include <riflib/types.hh>
 #include <riflib/scaffold/ScaffoldDataCache.hh>
 #include <riflib/rifdock_tasks/HackPackTasks.hh>
+#include <riflib/ScoreRotamerVsTarget.hh>
 
 #include <core/chemical/ChemicalManager.hh>
 #include <core/chemical/ResidueTypeSet.hh>
